+++ CHANGELOG.md
@@ -987,7 +987,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [@adguard/logger]: https://github.com/AdguardTeam/tsurlfilter/blob/master/packages/logger/CHANGELOG.md
 
 [Unreleased]: https://github.com/AdguardTeam/FiltersCompiler/compare/v3.4.0...HEAD
-[3.4.0]: https://github.com/AdguardTeam/FiltersCompiler/compare/v3.3.2-beta.0...v3.4.0
+[3.4.0]: https://github.com/AdguardTeam/FiltersCompiler/compare/v3.3.1...v3.4.0
 [3.3.2-beta.0]: https://github.com/AdguardTeam/FiltersCompiler/compare/v3.3.1...v3.3.2-beta.0
 [3.3.1]: https://github.com/AdguardTeam/FiltersCompiler/compare/v3.3.0...v3.3.1
 [3.3.0]: https://github.com/AdguardTeam/FiltersCompiler/compare/v3.3.0-beta.1...v3.3.0